Create an abstract base class \Drush\Boot\BaseBoot to hold the non-Drupal-specific bootstrap code. Also create \Drush\Boot\EmptyBoot, a placeholder bootstrap class used when there is no site root folder. Also moved the bootstrap code that is not Drupal-dependent from lib/Drush/Boot/bootstrap.inc to includes/bootstrap.inc.
