+++ includes/complete.inc
@@ -250,6 +250,7 @@ function drush_complete_process_argv() {
   // Check for and record any site alias.
   drush_sitealias_check_arg();
   drush_sitealias_check_site_env();
+  $bootstrap = drush_preflight_bootstrap_class();
 
   // Return the new argv for easy reference.
   return $argv;
